+++ ReaderCollection/app/src/main/java/aragones/sergio/readercollection/models/base/GoogleVolumeResponse.kt
@@ -30,7 +30,7 @@ data class GoogleVolumeResponse(
     @SerializedName("categories")
     val categories: List<String>?,
     @SerializedName("averageRating")
-    val averageRating: Int?,
+    val averageRating: Double?,
     @SerializedName("ratingsCount")
     val ratingsCount: Int?,
     @SerializedName("imageLinks")
